Move get_candidate_filename() to the parent class
